Move rest of tests out of StripeTest
